tests: reduce number of critical cucumber tests
