+++ common/src/configuration/network.rs
@@ -82,6 +82,7 @@ impl FromStr for Network {
             "dibbler" => Ok(Dibbler),
             invalid => Err(ConfigurationError::new(
                 "network",
+                Some(value.to_string()),
                 &format!("Invalid network option: {}", invalid),
             )),
         }
