+++ integration_tests/cucumber.js
@@ -1,6 +1,7 @@
 module.exports = {
   default:
     "--tags 'not @long-running and not @wallet-ffi and not @broken' --fail-fast",
-  ci: " ",
+  none: " ",
+  ci: "--tags '@critical and not @long-running and not @broken ' --fail-fast",
   critical: "--format @cucumber/pretty-formatter --tags @critical",
 };
